--- js/src/credential-form.js
@@ -1,10 +1,0 @@
-export default function CredentialForm() { }
-
-CredentialForm.show = function () {
-    document.getElementById("user-credential-form").hidden = false;
-}
-
-
-CredentialForm.hide = function () {
-    document.getElementById("user-credential-form").hidden = true;
-}
