+++ js/src/credential-formts.js
@@ -1,0 +1,9 @@
+export default class CredentialForm {
+    static show() {
+        document.getElementById("user-credential-form").hidden = false;
+    }
+    static hide() {
+        document.getElementById("user-credential-form").hidden = true;
+    }
+}
+//# sourceMappingURL=credential-formts.js.map
