--- js/src/load-icon.js
@@ -1,8 +1,0 @@
-export default function LoadIcon() { }
-
-LoadIcon.show = function () {
-    document.getElementById("loading-icon").hidden = false;
-}
-LoadIcon.hide = function () {
-    document.getElementById("loading-icon").hidden = true;
-}
