+++ js/src/load-iconts.js
@@ -1,0 +1,9 @@
+export default class LoadIcon {
+    static show() {
+        document.getElementById("loading-icon").hidden = false;
+    }
+    static hide() {
+        document.getElementById("loading-icon").hidden = true;
+    }
+}
+//# sourceMappingURL=load-iconts.js.map
