+++ js/ts/load-iconts.ts
@@ -1,0 +1,8 @@
+export default class LoadIcon {
+    static show() {
+        document.getElementById("loading-icon").hidden = false;
+    }
+    static hide() {
+        document.getElementById("loading-icon").hidden = true;
+    }
+}
