+++ npm-debug.log
@@ -1,0 +1,39 @@
+0 info it worked if it ends with ok
+1 verbose cli [ 'C:\\Program Files (x86)\\nodejs\\\\node.exe',
+1 verbose cli   'C:\\Program Files (x86)\\nodejs\\node_modules\\npm\\bin\\npm-cli.js',
+1 verbose cli   'run',
+1 verbose cli   'watch' ]
+2 info using npm@2.7.4
+3 info using node@v0.12.2
+4 verbose node symlink C:\Program Files (x86)\nodejs\\node.exe
+5 verbose run-script [ 'prewatch', 'watch', 'postwatch' ]
+6 info prewatch gitdone@1.0.0
+7 info watch gitdone@1.0.0
+8 verbose unsafe-perm in lifecycle true
+9 info gitdone@1.0.0 Failed to exec watch script
+10 verbose stack Error: gitdone@1.0.0 watch: `webpack --watch`
+10 verbose stack Exit status 1
+10 verbose stack     at EventEmitter.<anonymous> (C:\Program Files (x86)\nodejs\node_modules\npm\lib\utils\lifecycle.js:213:16)
+10 verbose stack     at EventEmitter.emit (events.js:110:17)
+10 verbose stack     at ChildProcess.<anonymous> (C:\Program Files (x86)\nodejs\node_modules\npm\lib\utils\spawn.js:14:12)
+10 verbose stack     at ChildProcess.emit (events.js:110:17)
+10 verbose stack     at maybeClose (child_process.js:1015:16)
+10 verbose stack     at Process.ChildProcess._handle.onexit (child_process.js:1087:5)
+11 verbose pkgid gitdone@1.0.0
+12 verbose cwd F:\dev\gitdone
+13 error Windows_NT 6.3.9600
+14 error argv "C:\\Program Files (x86)\\nodejs\\\\node.exe" "C:\\Program Files (x86)\\nodejs\\node_modules\\npm\\bin\\npm-cli.js" "run" "watch"
+15 error node v0.12.2
+16 error npm  v2.7.4
+17 error code ELIFECYCLE
+18 error gitdone@1.0.0 watch: `webpack --watch`
+18 error Exit status 1
+19 error Failed at the gitdone@1.0.0 watch script 'webpack --watch'.
+19 error This is most likely a problem with the gitdone package,
+19 error not with npm itself.
+19 error Tell the author that this fails on your system:
+19 error     webpack --watch
+19 error You can get their info via:
+19 error     npm owner ls gitdone
+19 error There is likely additional logging output above.
+20 verbose exit [ 1, true ]
