+++ package.json
@@ -7,10 +7,13 @@
         "webpack": "^3.10.0"
     },
     "devDependencies": {
+        "typescript": "^2.6.2",
         "clean-webpack-plugin": "^0.1.18"
     },
     "scripts": {
-        "test": "echo \"Error: no test specified\" && exit 1"
+        "test": "echo \"Error: no test specified\" && exit 1",
+        "build": "webpack",
+        "watch": "webpack --watch"
     },
     "repository": {
         "type": "git",
